Coverity fixes / Adjustment to way flightline works / OpenTarget Triangle moderator
--  Far to big a mod - adds fix for flightlines -- this still needs
simplifying
--  Covertity fixes
-- ESS BeRef fixes to make a lot quicker and more stable.
-- Addition of a open W block target
-- Partial update of triangle moderator to be non-triangle like if
needed

+++ chipInc/ChipIRInsert.h
@@ -89,7 +89,7 @@ class ChipIRInsert : public BulkInsert
 
  public:
 
-  ChipIRInsert(const int,const std::string&,const std::string&);
+  ChipIRInsert(const size_t,const std::string&,const std::string&);
   ChipIRInsert(const ChipIRInsert&);
   ChipIRInsert& operator=(const ChipIRInsert&);
   virtual ~ChipIRInsert();
